Fix async component update errors in tests
We have defined our own `waitFor` as a temporary work around for a bug
in RNTL. However, we should likely prefer the library's provided async
utils where possible. This leverages `findBy` queries which use
`waitFor` themselves.

Additionally, while `fireEvent` includes an `act` wrapper for async
component updates, it would appear that it does not catch updates in our
components triggered by the `onLayout` hook. Therefore, the `fireEvent`
is wrapped in an async `act` call to address the test errors.

Lastly, Testing Library provides and recommends leveraging the `screen`
export for queries. However, React Native Testing Library does not yet
support it. A recommeneded workaround is to name the return of render
`screen`.

- https://bit.ly/2WKK4t6
- https://git.io/Juvz5