Do not separately sanitize attributes in cloneSanitizedBlock
The only reason we still had this was to fill in default values for
any internal attrs that got stripped on duplication; since there is
no obvious use case for an internal attribute with defaults I'm
removing this for now.

I leave __experimentalSanitizeBlockAttributes as it is still used in
createBlock, and keeping it pulled out like this allows for ease of
unit testing.