Fix non-dismissible modals caused by header layout transitions
This change ensures changes to the header height are considered when
computing the allowable maximum height for the `BottomSheet`.

When the `BottomSheet` header changes its size, it configures a
`LayoutAnimation` to manage the change. It then relies upon the
`onSetHeight` callback to implement the change. In the scenario where
the _header_ layout changed, but the rest of the content did not, it
meant we attempted to set the height to the same value. This would
result in an unconsumed `LayoutAnimation` lingering as no changes
occurred which would consume the `LayoutAnimation`.

Unconsumed `LayoutAnimations` appear to cause issues for React Native's
Modal component on iOS. Specifically, it can result in a transparent,
non-dismissible modal sitting atop the rest of the app UI. Upgrading to
React Native 0.66 appears to have increased the frequency of this
occurring, for an unknown reason.

- https://git.io/J1W3z
- https://git.io/J1W32
- https://git.io/J1W3a
- https://git.io/J1W3r
- https://git.io/J1W36